SocialNetwork/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
environ==1.0
idna==3.3
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
